add update timing on myrewards
